Update and rename PurityWisp.cs to ExampleHoveringMinion.cs
